src/archivematica-s3-file-retrieval/mail_notification_api/requirements.txt
certifi==2022.5.18.1
python-jose==3.3.0
mysqlclient==2.1.0
fastapi==0.65.1
uvicorn==0.15.0
sqlalchemy==1.4.18
passlib==1.7.4
requests==2.27.1
bcrypt==3.2.2